fix: changed to tx
