+++ packages/app/config/index.ts
@@ -32,7 +32,6 @@ export default {
   socials,
   officialURL,
   openseaApi,
-  officialURL,
   chainIdMapping,
   zeroAddress,
   backendURL,
